requirements.txt
streamlit>=1.33.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
duckduckgo-search>=3.9.6
pillow>=10.3.0
pytest>=8.1.1